Use smooth-scrolling with hash links.
